WFLY-15838 Move our single Hibernate Search test to a dedicated package
So that we can add more tests in sibling packages.
