+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/hibernate/search/simple/HibernateSearchJPATestCase.java
@@ -19,7 +19,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.as.test.integration.hibernate.search;
+package org.jboss.as.test.integration.hibernate.search.simple;
 
 import static org.jboss.as.test.shared.PermissionUtils.createPermissionsXmlAsset;
 import static org.junit.Assert.assertEquals;
