+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/hibernate/search/simple/SearchBean.java
@@ -19,7 +19,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.as.test.integration.hibernate.search;
+package org.jboss.as.test.integration.hibernate.search.simple;
 
 import java.util.List;
 import javax.ejb.Stateful;
