+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/hibernate/search/simple/persistence.xml
